Move types to places where they are being usedCorrect typo - "description" instead of discription"
