+++ src/initials/initials.tsx
@@ -1,15 +1,15 @@
 export const today = new Date().toISOString().substring(0, 10);
 
 export const initialTodoForm = {
-    id: 0,
-    discription: "",
-    deadline: today,
-    important: false,
-    done: false,
-    doneDate: ""
+  id: 0,
+  description: "",
+  deadline: today,
+  important: false,
+  done: false,
+  doneDate: ""
 };
 
 export const initialState = {
-    todoForm: initialTodoForm,
-    todoList: []
+  todoForm: initialTodoForm,
+  todoList: []
 };
